requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
azure-common==1.1.18
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b3
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.4.0
azure-storage-nspkg==3.1.0
boto==2.49.0
boto3==1.9.131
botocore==1.12.131
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
docutils==0.14
funcy==1.11
future==0.17.1
gensim==3.5.0
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
more-itertools==7.0.0
nltk==3.3
numexpr==2.6.9
numpy==1.16.2
pandas==0.24.2
pluggy==0.9.0
protobuf==3.6.1
py==1.8.0
pycparser==2.19
pyLDAvis==2.1.2
pytest==4.4.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
six==1.11.0
smart-open==1.8.1
urllib3==1.24.1

